# swagger-spec.json
{"openapi":"3.0.0","paths":{"/account":{"get":{"operationId":"accountDetailsByClientId","summary":"Fetch account Details by Client Id","parameters":[{"name":"Version-Header","in":"header","description":"please select a version","required":true,"schema":{"enum":["1"],"type":"string"}},{"name":"clientId","required":true,"in":"query","schema":{"type":"string"}},{"name":"trackingId","required":true,"in":"query","schema":{"type":"string"}},{"name":"accountState","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"Account details fetched successfully","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AccountDetailsByClientIdResponse"}}}}},"400":{"description":"Invalid payload for Mambu"},"401":{"description":"Invalid username or password for Mambu"},"404":{"description":"Resource not found in Mambu"},"500":{"description":"Internal Server Error"}},"tags":["account-service"]}},"/account/account-numbers":{"get":{"operationId":"accountNumbersByClientId","summary":"Fetch account numbers by client Id","parameters":[{"name":"Version-Header","in":"header","description":"please select a version","required":true,"schema":{"enum":["1"],"type":"string"}},{"name":"externalCustomerId","required":true,"in":"query","schema":{"type":"string"}},{"name":"trackingId","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"Account numbers fetched successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccountNumbersByClientIdResponse"}}}},"400":{"description":"Invalid payload for Mambu"},"401":{"description":"Invalid username or password for Mambu"},"404":{"description":"Resource not found in Mambu"},"500":{"description":"Internal Server Error"}},"tags":["account-service"]}},"/account/{accountId}":{"get":{"operationId":"accountDetailsByAccountId","summary":"Fetch account Details by Account Id","parameters":[{"name":"Version-Header","in":"header","description":"please select a version","required":true,"schema":{"enum":["1"],"type":"string"}},{"name":"accountId","required":true,"in":"path","description":"The id or encoded key of the deposit account\t","schema":{"type":"string"}},{"name":"trackingId","required":true,"in":"query","description":"The trackingId of the transaction","schema":{"type":"string"}}],"responses":{"200":{"description":"Account details fetched successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccountDetailsByAccountIdResponse"}}}},"400":{"description":"Invalid payload for Mambu"},"401":{"description":"Invalid username or password for Mambu"},"404":{"description":"Resource not found in Mambu"},"500":{"description":"Internal Server Error"}},"tags":["account-service"]}},"/account/{accountId}/accumulators":{"get":{"operationId":"accountLimitsByAccountId","summary":"Fetch account limits and accumulator by Account Id","parameters":[{"name":"Version-Header","in":"header","description":"please select a version","required":true,"schema":{"enum":["1"],"type":"string"}}],"responses":{"200":{"description":"Account limits fetched successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccountDetailsByAccountIdResponse"}}}},"400":{"description":"Invalid payload for PTS"},"401":{"description":"Invalid username or password for PTS"},"404":{"description":"Resource not found in PTS"},"500":{"description":"Internal Server Error"}},"tags":["account-service"]}},"/account/users/{userId}/account-numbers":{"get":{"operationId":"getAccountsByUserId","parameters":[{"name":"Version-Header","in":"header","description":"please select a version","required":true,"schema":{"enum":["1"],"type":"string"}}],"responses":{"200":{"description":""}},"tags":["account-service"]}},"/account/certificate/url":{"post":{"operationId":"getCertificate","summary":"Fetch certificate url from AWS S3 by templateName","parameters":[{"name":"Version-Header","in":"header","description":"please select a version","required":true,"schema":{"enum":["1"],"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetCertificateStrategyDataInputDto"}}}},"responses":{"200":{"description":"Certificate fetched successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetCertificateResponseDto"}}}}},"tags":["account-service"]}},"/account/modify-limits":{"post":{"operationId":"modifyLimits","summary":"Modify user account limits by accountNumber","parameters":[{"name":"Version-Header","in":"header","description":"please select a version","required":true,"schema":{"enum":["1"],"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ModifyLimitsInputDto"}}}},"responses":{"200":{"description":"Modify user account limits successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccountModifyLimitsResponse"}}}}},"tags":["account-service"]}},"/account/api/sendSMS":{"get":{"operationId":"consultElectronicDeposit","summary":"Account back office operations consult the electronic deposit","parameters":[{"name":"Version-Header","in":"header","description":"please select a version","required":true,"schema":{"enum":["1"],"type":"string"}},{"name":"phone","required":false,"in":"query","example":"3216545987","description":"Client mobile phone","schema":{"type":"number"}},{"name":"depositNumber","required":false,"in":"query","example":"2000004","description":"Client deposit number / mambu account id","schema":{"type":"number"}},{"name":"identification","required":false,"in":"query","example":"22412468","description":"Client identification number","schema":{"type":"number"}},{"name":"cambioEstado","required":false,"in":"query","example":"22412468","description":"Client identification number","schema":{"type":"string"}}],"responses":{"200":{"description":"Account back office operations consult the electronic deposit successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccountDetailsByAccountIdResponse"}}}},"400":{"description":"Invalid payload Account back office operations consult the electronic deposit"},"401":{"description":"Invalid Account back office operations consult the electronic deposit"},"404":{"description":"Resource not found in Account back office operations consult the electronic deposit"},"500":{"description":"Internal Server Error"}},"tags":["account-service"]}},"/service/config":{"get":{"operationId":"getConfig","parameters":[],"responses":{"200":{"description":""}}}},"/service/health":{"get":{"operationId":"check","parameters":[],"responses":{"200":{"description":""}}}}},"info":{"title":"account-service service","description":"The account-service service API descriptions","version":"1.0","contact":{}},"tags":[{"name":"account-service","description":""}],"servers":[],"components":{"schemas":{"AccountDetailsByClientIdResponse":{"type":"object","properties":{}},"AccountNumbersByClientIdResponse":{"type":"object","properties":{"accountNumbers":{"type":"array","items":{"type":"string"}}},"required":["accountNumbers"]},"AccountDetailsByAccountIdResponse":{"type":"object","properties":{"clientId":{"type":"string"}},"required":["clientId"]},"GetCertificateStrategyDataInputDto":{"type":"object","properties":{}},"GetCertificateResponseDto":{"type":"object","properties":{}},"ModifyLimitsInputDto":{"type":"object","properties":{"accountId":{"type":"string"}},"required":["accountId"]},"AccountModifyLimitsResponse":{"type":"object","properties":{}},"MessageEvent":{"type":"object","properties":{}},"AccountPTSEventDto":{"type":"object","properties":{"userId":{"type":"string"},"bPartnerId":{"type":"string"},"customerExternalId":{"type":"string"},"customerExternalNumber":{"type":"string"},"phoneNumber":{"type":"string"},"enrollmentId":{"type":"string"},"phonePrefix":{"type":"string"}},"required":["userId","bPartnerId","customerExternalId","customerExternalNumber","phoneNumber","enrollmentId","phonePrefix"]},"AccountEventDto":{"type":"object","properties":{"userId":{"type":"string"},"accountId":{"type":"string"},"phoneNumber":{"type":"string"},"enrollmentId":{"type":"string"},"phonePrefix":{"type":"string"},"customerExternalId":{"type":"string"},"customerExternalNumber":{"type":"string"}},"required":["userId","accountId","phoneNumber","enrollmentId","phonePrefix","customerExternalId","customerExternalNumber"]},"UpdateAccountEventDto":{"type":"object","properties":{"userId":{"type":"string"},"accountId":{"type":"string"},"crmDepositId":{"type":"string"},"crmContactAgreementId":{"type":"string"}},"required":["userId","accountId","crmDepositId","crmContactAgreementId"]}}}}